에코서버 모든오류 해결
원인은 %d 로 숫자출력하는 것과 (int) 안붙인거

+++ 현호의서버연습/ECHOclient/ECHOclient/source.cpp
@@ -58,14 +58,14 @@ int main()
 	{
 		sendData[0] = count;
 		send(hSocket, sendData, 1, 0);
-		cout << "send data : " << sendData[0] << endl;
+		cout << "send data : " << (int)sendData[0] << endl;
 		recvSize = recv(hSocket, recvData, sizeof(recvData), 0);
 		if (recvSize == SOCKET_ERROR)
 		{
 			cout << "recv() Error" << endl;
 			break;
 		}
-		cout << "recv Data : " << recvData[0] << endl;
+		cout << "recv Data : " << (int)recvData[0] << endl;
 	}
 
 	// 5. 소켓 종료
